requirements.txt
dj-database-url==0.3.0
Django==1.9.1
django-filter==0.12.0
djangorestframework==3.3.2
psycopg2==2.6.1
